+++ packages/ice/src/createService.ts
@@ -38,6 +38,7 @@ import ServerRunner from './service/ServerRunner.js';
 import RouteManifest from './utils/routeManifest.js';
 import dynamicImport from './utils/dynamicImport.js';
 import mergeTaskConfig from './utils/mergeTaskConfig.js';
+import addPolyfills from './utils/runtimePolyfill.js';
 
 const require = createRequire(import.meta.url);
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
@@ -163,10 +164,6 @@ async function createService({ rootDir, command, commandArgs }: CreateServiceOpt
   RUNTIME_EXPORTS.forEach(exports => {
     generatorAPI.addExport(exports);
   });
-  // Add polyfills.
-  generatorAPI.addEntryImportAhead({
-    source: '@ice/runtime/polyfills/signal',
-  });
   const routeManifest = new RouteManifest();
   const ctx = new Context<Config, ExtendsPluginAPI>({
     rootDir,
@@ -244,6 +241,8 @@ async function createService({ rootDir, command, commandArgs }: CreateServiceOpt
       },
     });
   }
+
+  addPolyfills(generatorAPI, userConfig.featurePolyfill, rootDir, command === 'start');
 
   // Get first task config as default platform config.
   const platformTaskConfig = taskConfigs[0];