+++ packages/runtime/src/index.ts
@@ -8,6 +8,7 @@ import {
 import type {
   RuntimePlugin,
   AppContext,
+  PublicAppContext,
   AppConfig,
   RouteConfig,
   RouteItem,
@@ -22,7 +23,7 @@ import type {
 import Runtime from './runtime.js';
 import runClientApp from './runClientApp.js';
 import type { RunClientAppOptions } from './runClientApp.js';
-import { useAppContext, useAppData, AppContextProvider } from './AppContext.js';
+import { useAppContext as useInternalAppContext, useAppData, AppContextProvider } from './AppContext.js';
 import { getAppData } from './appData.js';
 import { useData, useConfig } from './RouteContext.js';
 import {
@@ -32,6 +33,7 @@ import {
   Scripts,
   Main,
   Data,
+  usePageAssets,
 } from './Document.js';
 import type {
   MetaType,
@@ -53,6 +55,26 @@ import usePageLifecycle from './usePageLifecycle.js';
 import { withSuspense, useSuspenseData } from './Suspense.js';
 import { createRouteLoader, WrapRouteComponent, RouteErrorComponent, Await } from './routes.js';
 
+function useAppContext(): PublicAppContext {
+  const context = useInternalAppContext();
+
+  const {
+    appConfig,
+    routePath,
+    downgrade,
+    documentOnly,
+    renderMode,
+  } = context;
+
+  return {
+    appConfig,
+    routePath,
+    downgrade,
+    documentOnly,
+    renderMode,
+  };
+}
+
 export {
   getAppConfig,
   defineAppConfig,
@@ -73,6 +95,7 @@ export {
   Scripts,
   Data,
   Main,
+  usePageAssets,
   // API for data-loader.
   dataLoader,
   callDataLoader,