+++ packages/webpack-config/src/index.ts
@@ -21,6 +21,7 @@ import getSplitChunksConfig, { FRAMEWORK_BUNDLES } from './config/splitChunks.js
 import compilationPlugin from './unPlugins/compilation.js';
 import compileExcludes from './compileExcludes.js';
 
+export { isSupportedFeature } from './unPlugins/compilation.js';
 export { getCSSModuleLocalIdent } from './utils/getCSSModuleLocalIdent.js';
 
 const require = createRequire(import.meta.url);